+++ app/policies/account/emails_policy.rb
@@ -1,9 +1,8 @@
-class Account::EmailPasswordsPolicy < BasePolicy
+class Account::EmailsPolicy < BasePolicy
   def show?
     current_user.present?
   end
   alias_method :update_email?, :show?
-  alias_method :update_password?, :show?
   alias_method :resend_email_change?, :show?
   alias_method :cancel_email_change?, :show?
 end
