+++ app/policies/account/passwords_policy.rb
@@ -1,0 +1,6 @@
+class Account::PasswordsPolicy < BasePolicy
+  def show?
+    current_user.present?
+  end
+  alias_method :update_password?, :show?
+end
