+++ app/views/account/emails/show.html.erb
@@ -1,4 +1,4 @@
-<% content_for :title, "Edit your account" %>
+<% content_for :title, "Change your email address" %>
 
 <% content_for :breadcrumbs,
    render("govuk_publishing_components/components/breadcrumbs", {
@@ -13,7 +13,7 @@
          url: account_path,
        },
        {
-         title: "Edit your account",
+         title: "Change your email address",
        }
      ]
    })
@@ -26,8 +26,8 @@
   } do %>
     <p class="govuk-body">We've sent an email to <strong><%= current_user.unconfirmed_email %></strong> with a link to confirm the change. If you haven't received this email, we can send it again:</p>
 
-    <%= link_to "Resend confirmation email", resend_email_change_account_email_password_path, method: :put, class: "govuk-button app-button--no-margin" %>
-    <%= link_to "Cancel change", cancel_email_change_account_email_password_path, method: :delete, class: "govuk-link app-link--inline" %>
+    <%= link_to "Resend confirmation email", resend_email_change_account_email_path, method: :put, class: "govuk-button app-button--no-margin" %>
+    <%= link_to "Cancel change", cancel_email_change_account_email_path, method: :delete, class: "govuk-link app-link--inline" %>
   <% end %>
 <% end %>
 
@@ -47,9 +47,7 @@
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <h2 class="govuk-heading-m">Change your email address</h2>
-
-    <%= form_for current_user, url: update_email_account_email_password_path do |f| %>
+    <%= form_for current_user, url: update_email_account_email_path do |f| %>
       <%= render "govuk_publishing_components/components/input", {
         label: {
           text: "Email address"
@@ -64,13 +62,6 @@
       <%= render "govuk_publishing_components/components/button", {
         text: "Change email"
       } %>
-    <% end %>
-
-    <hr class="govuk-section-break govuk-section-break--visible govuk-section-break--xl">
-
-    <h2 class="govuk-heading-m">Change your password</h2>
-    <%= form_for current_user, :url => update_password_account_email_password_path do |f| %>
-      <%= render partial: "devise/passwords/change_password_panel", locals: { f: f, user: current_user, updating_password: true } %>
     <% end %>
   </div>
 </div>
