+++ app/views/accounts/show.html.erb
@@ -19,10 +19,15 @@
   items: [
     {
       link: {
-        text: "Change your email or password",
-        path: account_email_password_path,
+        text: "Change your email address",
+        path: account_email_path,
       },
-      description: "Reset your password or change your email address.",
+    },
+    {
+      link: {
+        text: "Change your password",
+        path: account_password_path,
+      },
     },
     (
       {
