+++ test/integration/account_test.rb
@@ -18,7 +18,8 @@ class AccountTest < ActionDispatch::IntegrationTest
       assert_current_url account_path
       assert page.has_selector?("h1", text: "Settings")
 
-      assert page.has_link?("Change your email or password", href: account_email_password_path)
+      assert page.has_link?("Change your email address", href: account_email_path)
+      assert page.has_link?("Change your password", href: account_password_path)
       assert page.has_link?("Manage permissions", href: account_manage_permissions_path)
       assert page.has_link?("Change your 2-step verification phone", href: two_step_verification_path)
       assert page.has_link?("View your role and organisation", href: account_role_organisation_path)
@@ -34,7 +35,8 @@ class AccountTest < ActionDispatch::IntegrationTest
       visit account_path
       assert page.has_selector?("h1", text: "Settings")
 
-      assert page.has_link?("Change your email or password", href: account_email_password_path)
+      assert page.has_link?("Change your email address", href: account_email_path)
+      assert page.has_link?("Change your password", href: account_password_path)
       assert page.has_link?("Change your 2-step verification phone", href: two_step_verification_path)
       assert page.has_link?("View your role and organisation", href: account_role_organisation_path)
       assert page.has_link?("Your account access log", href: account_activity_path)
