+++ test/support/password_helpers.rb
@@ -2,7 +2,7 @@ module PasswordHelpers
   def change_password(options)
     # this method assumes that the user is logged in
     visit root_path
-    click_link "Change your email or password"
+    click_link "Change your password"
     fill_in "Current password", with: options[:old]
     fill_in "New password", with: options[:new]
     fill_in "Confirm new password", with: options[:new_confirmation]
